Making changes to the controllers adding a new endpoint and fixing some warnings. WIP Testing

+++ Movies App/Movies.Api/Controllers/MoviesController.cs
@@ -28,7 +28,6 @@ namespace Movies.Api.Controllers
         [HttpPost(ApiEndpoints.Movies.Create)]
         public async Task<ResponseModel<string>> Create([FromBody] CreateMovieRequest request, CancellationToken token)
         {
-
             var response = new ResponseModel<string>
             {
                 Title = "Something went wrong.",
@@ -41,10 +40,12 @@ namespace Movies.Api.Controllers
             var result = await _movieService.CreateAsync(movie, token);
             response.Success = result.Success;
             response.Title = result.Title;
+            response.Content = result.Content;
 
             return response;
 
         }
+
         [AllowAnonymous]
         [HttpGet(ApiEndpoints.Movies.Get)]
         public async Task<ResponseModel<MovieResponse>> Get(Guid id, CancellationToken token)
@@ -58,7 +59,7 @@ namespace Movies.Api.Controllers
             try
             {
                 bool isAdmin = false;
-                string userId = null;
+                string userId = null!;
                 bool isAuthenticated = HttpContext.IsUserAuthenticated(out var authenticatedUserId);
 
                 if (isAuthenticated)
@@ -228,26 +229,37 @@ namespace Movies.Api.Controllers
                 Title = "Something went wrong.",
                 Success = false
             };
+
             try
             {
                 var userId = HttpContext.GetUserId();
-                var movie = request.MapToMovie(request.Id);
-                var updatedMovie = await _movieService.UpdateAsync(movie, userId, token);
+                var isAdmin = HttpContext.CheckAdmin();
 
-                if (updatedMovie != null)
+                if (isAdmin)
                 {
-                    var movieDetail = updatedMovie.MapToResponse();
-                    response.Title = "The title has been successfully updated.";
-                    response.Success = true;
-                    response.Content = movieDetail;
+                    var movie = request.MapToMovie(request.Id, userId.ToString());
+                    var updatedMovie = await _movieService.UpdateAsync(movie, userId, token);
+
+                    if (updatedMovie.Success)
+                    {
+                        var movieDetail = updatedMovie.Content.MapToResponse();
+                        response.Title = "Title updated successfully";
+                        response.Success = updatedMovie.Success;
+                        response.Content = movieDetail;
+                    }
                 }
+                else
+                {
+                    response.Success = false;
+                    response.Title = "Unable to update title: Unauthorized access.";
+                }
+
             }
             catch (Exception ex)
             {
                 response.Success = false;
                 response.Title = ex.Message;
             }
-
 
             return response;
         }
@@ -261,23 +273,35 @@ namespace Movies.Api.Controllers
                 Title = "Something went wrong.",
                 Success = false
             };
+
             try
             {
-                var deleted = await _movieService.DeleteByIdAsync(id, token);
-                if (deleted)
-                {
-                    response.Success = true;
-                    response.Title = "Title deleted successfully.";
-                }
+                bool isAdmin = HttpContext.CheckAdmin();
 
+                if (isAdmin)
+                {
+                    var deleted = await _movieService.DeleteByIdAsync(id, token);
+                    if (deleted)
+                    {
+                        response.Success = true;
+                        response.Title = "Title successfully deleted.";
+                    }
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Title = "Unable to delete movie: Unauthorized access.";
+                }
             }
             catch (Exception ex)
             {
                 response.Success = false;
                 response.Title = ex.Message;
             }
+
             return response;
         }
+
         [AllowAnonymous]
         [HttpGet(ApiEndpoints.Movies.Search)]
         public async Task<ResponseModel<IEnumerable<Movie>>> SearchMovie(string textToSearchMovie)
@@ -291,6 +315,7 @@ namespace Movies.Api.Controllers
             try
             {
                 var moviesList = await _movieService.GetSearchedMoviesAsync(textToSearchMovie);
+
                 if (moviesList != null)
                 {
                     response.Content = moviesList;
@@ -304,7 +329,6 @@ namespace Movies.Api.Controllers
                 response.Title = ex.Message;
             }
 
-
             return response;
         }
 
@@ -316,6 +340,7 @@ namespace Movies.Api.Controllers
                 Title = "Something went wrong.",
                 Success = false
             };
+
             try
             {
                 if (!Guid.TryParse(movieId, out var guidMovieId))
@@ -338,8 +363,6 @@ namespace Movies.Api.Controllers
 
             return response;
         }
-
-
 
         [HttpDelete(ApiEndpoints.Movies.RemoveFromWatchlist)]
         public async Task<ResponseModel<string>> DeleteMovieFromWatchList(Guid userWatchlistId)
@@ -382,8 +405,7 @@ namespace Movies.Api.Controllers
             {
                 var userId = HttpContext.GetUserId();
                 var isAdmin = HttpContext.CheckAdmin();
-                var allWatchlistData =
-                await _userWatchlistService.GetAllAsync(userId.ToString(), isAdmin);
+                var allWatchlistData = await _userWatchlistService.GetAllAsync(userId.ToString(), isAdmin);
                 var result = allWatchlistData.MapToResponse(1, 10, 10);
 
                 if (allWatchlistData != null)
@@ -404,7 +426,6 @@ namespace Movies.Api.Controllers
 
         [AllowAnonymous]
         [HttpGet(ApiEndpoints.Movies.GetTopMovies)]
-
         public async Task<ResponseModel<MoviesResponse>> TopMoviesList(CancellationToken token)
         {
             ResponseModel<MoviesResponse> response = new ResponseModel<MoviesResponse>
@@ -425,7 +446,7 @@ namespace Movies.Api.Controllers
 
                 var result = isUserAuthenticated
                     ? await _movieService.GetTopMovieAsync(isAdmin, userId.ToString(), token)
-                    : await _movieService.GetTopMovieAsync(false, null, token);
+                    : await _movieService.GetTopMovieAsync(false, null!, token);
 
                 if (result.Success)
                 {
@@ -500,7 +521,7 @@ namespace Movies.Api.Controllers
             try
             {
                 bool isAdmin = false;
-                string userId = null;
+                string userId = null!;
                 bool isAuthenticated = HttpContext.IsUserAuthenticated(out var authenticatedUserId);
 
                 if (isAuthenticated)
